+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.0] - 2025-11-14
+### Added
+- Make subscription payment
+- Make order payment
+
 ## [1.1.2] - 2025-04-16
 ### Added
 - added license for order products and subscription